Factor sticker index/send into separate routes
