+++ app/router.js
@@ -18,7 +18,11 @@ AppRouter.map(function() {
     this.route('login');
     this.route('register');
   });
-  this.route('app', function() {});
+  this.route('app', function() {
+    this.route('stickers', function() {
+      this.route('send');
+    });
+  });
 });
 
 export default AppRouter;
