+++ app/templates/app/index.hbs
@@ -7,48 +7,5 @@
     {{#rev-user-info as |user|}}
       Thanks for logging in <b>{{user.email}}</b>
     {{/rev-user-info}}
-
-    <div class="mt4 mb4">
-    <h3 class="f3 mb3 tc baskerville">Stickers</h3>
-    {{#each model.stickers as |sticker|}}
-
-    {{else}}
-      <span class="db tc">
-        No stickers yet ☹️
-      </span>
-    {{/each}}
-    </div>
-
-  {{/rev-card}}
-
-  {{#rev-card
-    buttonText='Send Sticker'
-    onsubmit=(route-action 'createSticker')
-    title='Send a sticker!'}}
-
-  <div class="mb3">
-    {{rev-input
-      value=model.newSticker.receiver
-      classNames=''
-      type='email'
-      label='Receiver email'}}
-  </div>
-
-  <div class="mb3">
-    {{rev-input
-      value=model.newSticker.title
-      classNames=''
-      type=''
-      label='Title'}}
-  </div>
-
-  {{#if model.newSticker.errors}}
-    <div class="mb3">
-      {{#each model.newSticker.errors as |error|}}
-        <span class="db red mb2">{{error}}</span>
-      {{/each}}
-    </div>
-  {{/if}}
-
   {{/rev-card}}
 </div>
