+++ app/templates/app/stickers.hbs
@@ -1,0 +1,1 @@
+{{outlet}}
