+++ config/environment.js
@@ -56,7 +56,7 @@ module.exports = function(environment) {
     // ENV.APP.LOG_TRANSITIONS = true;
     // ENV.APP.LOG_TRANSITIONS_INTERNAL = true;
     // ENV.APP.LOG_VIEW_LOOKUPS = true;
-
+    ENV.enabled = false;
     ENV.serviceWorker.debug = false;
   }
